YoutubeUploaderApp/requirements.txt
Django==5.2.7
google-auth-oauthlib==1.2.0
google-api-python-client==2.123.0
google-auth==2.28.0
Pillow==10.2.0
python-dotenv==1.0.0
ffmpeg-python==0.2.0
